refactor Math.kt
