+++ app/src/main/java/com/example/tonetuner_v2/extensions/CollectionExtensions.kt
@@ -1,0 +1,8 @@
+package com.example.tonetuner_v2.extensions
+
+import kotlin.math.roundToInt
+
+fun <T> Collection<T>.median(): T where T : Comparable<T> {
+    val midIndex = (this.size / 2f).roundToInt()
+    return this.toList().sorted()[midIndex]
+}
