+++ app/src/main/java/com/example/tonetuner_v2/extensions/IntExtensions.kt
@@ -1,0 +1,4 @@
+package com.example.tonetuner_v2.extensions
+
+fun Int.factorial() =
+    (1..this).reduce { acc, i -> acc * i }
