+++ app/src/main/java/com/example/tonetuner_v2/ui/composables/FftOrSpectrumViewer.kt
@@ -12,9 +12,9 @@ import com.example.tonetuner_v2.app.AppModel
 import com.example.tonetuner_v2.app.AppModel.spectrumType
 import com.example.tonetuner_v2.audio.audioProcessing.toFingerPrint
 import com.example.tonetuner_v2.audio.audioProcessing.toGraphRepr
+import com.example.tonetuner_v2.extensions.step
 import com.example.tonetuner_v2.ui.composables.old.XYPlot
 import com.example.tonetuner_v2.ui.navigation.MainLayout
-import com.example.tonetuner_v2.util.step
 
 // todo Composable is too stateful.
 // todo fft does not plot correctly
