+++ app/src/main/java/com/example/tonetuner_v2/util/Math.kt
@@ -1,14 +1,11 @@
 package com.example.tonetuner_v2.util
 
 import com.example.tonetuner_v2.audio.audioProcessing.Harmonic
-import java.math.RoundingMode
-import java.text.DecimalFormat
+import com.example.tonetuner_v2.extensions.factorial
 import kotlin.math.pow
-import kotlin.math.roundToInt
 
-fun percentage(value: Number, total: Number): Float {
-    return (value.toFloat() / total.toFloat()) * 100f
-}
+fun percentage(value: Number, total: Number): Float =
+    (value.toFloat() / total.toFloat()) * 100f
 
 fun calcError(expected: Number, actual: Number): Float {
     expected as Float
@@ -16,54 +13,20 @@ fun calcError(expected: Number, actual: Number): Float {
     return ((actual - expected) / expected) * 100
 }
 
-fun <T> Collection<T>.median(): T where T : Comparable<T> {
-    val midIndex = (this.size / 2f).roundToInt()
-    return this.toList().sorted()[midIndex]
-}
-
-fun Int.fact() =
-    (1..this).reduce { acc, i -> acc * i }
-
 fun combinations(n: Int, k: Int) =
-    n.fact() / k.fact() * (n - k).fact()
-
-infix fun ClosedRange<Float>.step(step: Float): Iterable<Float> {
-    require(start.isFinite())
-    require(endInclusive.isFinite())
-    require(step > 0.0) { "Step must be positive, was: $step." }
-    val sequence = generateSequence(start) { previous ->
-        if (previous == Float.POSITIVE_INFINITY) return@generateSequence null
-        val next = previous + step
-        if (next > endInclusive) null else next
-    }
-    return sequence.asIterable()
-}
-
-fun Float.toRadian() = this * Math.PI.toFloat() / 180
-
-fun Float.toDegree() = this * 180 / Math.PI.toFloat()
-
-fun List<Float>.normalizeBySum(): List<Float> {
-    val sum = this.sum()
-    return this.map { it / sum }
-}
-
-@JvmName("sumListsFloat")
-fun List<List<Float>>.sumLists(): List<Float> =
-    this.groupByIndex().map { it.sum() }
+    n.factorial() / k.factorial() * (n - k).factorial()
 
 fun poly(x: List<Float>, y: List<Float>): Harmonic {
-    val coef = polyFit(x, y)
-    val a = coef[0]
-    val b = coef[1]
-    val c = coef[2]
-
-    return Harmonic(-b / (2 * a), c - b.pow(2) / (4 * a))
+    val (a, b, c) = polyFit(x, y)
+    return Harmonic(
+        freq = -b / (2 * a),
+        mag = c - b.pow(2) / (4 * a)
+    )
 }
 
-fun quadInterp(x: Float, xVals: List<Float>, yVals: List<Float>): Float {
-    val coef = polyFit(xVals, yVals)
-    return coef[0] * x.pow(2) + coef[1] * x + coef[2]
+fun quadInterp(x: Float, xAxis: List<Float>, yAxis: List<Float>): Float {
+    val (a, b, c) = polyFit(xAxis, yAxis)
+    return a * x.pow(2) + b * x + c
 }
 
 fun polyFit(x: List<Float>, y: List<Float>): List<Float> {
@@ -75,36 +38,26 @@ fun polyFit(x: List<Float>, y: List<Float>): List<Float> {
     return listOf(a, b, c)
 }
 
-fun List<Float>.normalize(
-    lowerBound: Float = -1f,
-    upperBound: Float = 1f
-) = this.toMutableList().apply { normalize(lowerBound, upperBound) }
+@JvmInline
+value class Degree(val value: Float) {
+    fun toRadians(): Radian =
+        (value * Math.PI.toFloat() / 180).radians
 
-fun MutableList<Float>.normalize(
-    lowerBound: Float = -1f,
-    upperBound: Float = 1f
-) {
-    // Check that array isn't empty
-    if (isEmpty()) return
-
-    val minValue = this.minByOrNull { it }!!
-    val maxValue = this.maxByOrNull { it }!!
-    val valueRange = (maxValue - minValue).toFloat()
-    val boundRange = (upperBound - lowerBound).toFloat()
-
-    // Check that array isn't already normalized
-    // (I would use in range, but this produces excess memory)
-    if ((minValue == 0f && maxValue == 0f) ||
-        (
-            maxValue <= upperBound && maxValue > upperBound &&
-                minValue >= lowerBound && minValue < lowerBound
-            )
-    ) {
-        return
-    }
-
-    // Normalize
-    for (i in indices) {
-        this[i] = ((boundRange * (this[i] - minValue)) / valueRange) + lowerBound
-    }
+    override fun toString(): String =
+        "$value°"
 }
+
+@JvmInline
+value class Radian(val value: Float) {
+    fun toDegrees(): Degree =
+        (value * 180 / Math.PI.toFloat()).degrees
+
+    override fun toString(): String =
+        "$value rad"
+}
+
+val Float.degrees: Degree
+    get() = Degree(this)
+
+val Float.radians: Radian
+    get() = Radian(this)
